Moved server_thread to javalinker-module + redesigned layout of javalinker
