+++ ROS/workspace/src/corelinker/scripts/handlers/java_module.py
@@ -1,9 +1,6 @@
-from location import Location
-
-import socket
 import json
+import socket
 import time
-from threading import Thread
 
 TCP_IP = '127.0.0.1'
 TCP_PORT_JAVA_PYTH = 5005
@@ -53,37 +50,6 @@ def connect():
     logger.log_info("[JAVAMODULE] Connected to Core")
     global connected
     connected = True
-
-
-class ServerThread(Thread):
-    def __init__(self, _IP_, _PORT_, _BUFFER_):
-        Thread.__init__(self)
-        self._IP_ = _IP_
-        self._PORT_ = _PORT_
-        self._BUFFER_ = _BUFFER_
-
-        self.server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.server_socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        self.server_socket.bind((self._IP_, self._PORT_))
-
-    def run(self):
-        self.server_socket.listen(4)
-        while True:
-            (conn, (ip, port)) = self.server_socket.accept()
-            data = read_line(conn)
-            data_string = str("".join(data))
-            logger.log_debug("[SERVERSOCKET] Server received data: " + data_string)
-            json_string = json.loads(data_string)
-
-            # from scripts import javalinker
-            # javalinker.get_type(json_string)
-            logger.log_info(__name__)
-
-
-def start_thread():
-    newthread = ServerThread(TCP_IP, TCP_PORT_JAVA_PYTH, BUFFER_SIZE)
-    newthread.daemon = True
-    newthread.start()
 
 
 def set_logger(logger_argument):
